Extract common list/item/checkbox component and add lineThrough feature
