+++ src/components/organisms/recipe/detail/ingredientsList/components/item.jsx
@@ -1,16 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import {
-  ListItem as Item,
-  ListItemIcon as Icon,
-  ListItemText as Text,
-  ListItemSecondaryAction as Action,
-} from '@material-ui/core';
+import { ListItem as Item, ListItemSecondaryAction as Action } from '@material-ui/core';
 import CommentIcon from '@material-ui/icons/Comment';
 
 import IconButton from '../../../../../atoms/buttons/icon.jsx';
-import Checkbox from '../../../../../atoms/fields/checkbox.jsx';
+import Checkbox from '../../../../list/item/checkbox.jsx';
 
 const Component = ({ name, emoji, quantity, isOptional, alternatives, value, handleClick }) => {
   let label = emoji ? `${emoji} (${name})` : name;
@@ -18,10 +13,7 @@ const Component = ({ name, emoji, quantity, isOptional, alternatives, value, han
 
   return (
     <Item role={undefined} dense button onClick={handleClick}>
-      <Icon>
-        <Checkbox value={value} />
-      </Icon>
-      <Text primary={isOptional ? `{${label}}` : `${label}`} />
+      <Checkbox value={value} text={isOptional ? `{${label}}` : `${label}`} />
       {alternatives && alternatives.length > 0 && (
         <Action>
           <IconButton edge="end" aria-label="comments">
