+++ src/components/organisms/recipe/detail/stepsList/components/item.jsx
@@ -1,17 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { ListItem as Item, ListItemIcon as Icon, ListItemText as Text } from '@material-ui/core';
+import { ListItem as Item } from '@material-ui/core';
 
-import Checkbox from '../../../../../atoms/fields/checkbox.jsx';
+import Checkbox from '../../../../list/item/checkbox.jsx';
 import MarkdownText from '../../../../../atoms/markdownText.jsx';
 
 const Component = ({ index, rowCounter, step, value, handleClick }) => (
   <Item role={undefined} dense button onClick={handleClick}>
-    <Icon>
-      <Checkbox value={value} />
-    </Icon>
-    <Text primary={<MarkdownText id={`${index}-${rowCounter}`} text={step} />} />
+    <Checkbox value={value} text={<MarkdownText id={`${index}-${rowCounter}`} text={step} />} />
   </Item>
 );
